prep for re-using the async io code
